Move the global tracing into detail from transforms
